Avoid boot delays caused by interfaces
Having the interfaces configured as auto will delay boot for more than 1
minute while it waits for the interfaces to be connected if they are not
connected - really long timeouts.

Having just allow-hotplug will only try to enable the interfaces once
udev identifies them. Still allowing them to be configured on boot - but
asynchronously - which won't impact boot time that much.

Signed-off-by: Daniel Finimundi <daniel@finimundi.com>